+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/TargetAssetAllocationsTest.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.databind.exc.ValueInstantiationException;
 import com.strangegrotto.wealthdraft.Main;
 import com.strangegrotto.wealthdraft.assetallocation.ExpectedExampleTargetAssetAllocations;
 import com.strangegrotto.wealthdraft.assetallocation.TargetAssetAllocationsTestFiles;
+import com.strangegrotto.wealthdraft.assetfilters.AssetFilter;
+import com.strangegrotto.wealthdraft.assetfilters.FiltersTestFiles;
 import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitions;
 import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitionsTestFiles;
 import com.strangegrotto.wealthdraft.assets.definition.ExpectedExampleAssetDefinitions;
@@ -11,6 +13,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.Map;
 
 public class TargetAssetAllocationsTest {
@@ -71,6 +74,11 @@ public class TargetAssetAllocationsTest {
         var assetsUrl = AssetDefinitionsTestFiles.EXAMPLE.getResource();
         var assetDefinitions = mapper.readValue(assetsUrl, AssetDefinitions.class);
         Main.addDeserializersNeedingAssetDefs(mapper, assetDefinitions);
+
+        var filtersUrl = FiltersTestFiles.EXAMPLE.getResource();
+        var mapType = mapper.getTypeFactory().constructMapType(HashMap.class, String.class, AssetFilter.class);
+        Map<String, AssetFilter> filters = mapper.readValue(filtersUrl, mapType);
+        Main.addDeserializersNeedingFilters(mapper, filters);
 
         var assetAllocationsUrl = testFile.getResource();
         return mapper.readValue(assetAllocationsUrl, TargetAssetAllocations.class);