+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## TBD
+* Split asset filters into their own file
+
 ## 0.13.0
 * Remove negation filter now that default tags are available, to guarantee that filters strictly reduce the number of assets
 * Add a new asset filter type, `filter`, which allows users to compose filters of other filters